requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# FastAPI + WebSocket server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9

# Audio I/O (used by legacy CLI main.py)
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Utilities
requests>=2.31.0
aiortc>=1.14.0
